pyproject.toml
[project]
name = "soloprojectfastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Takash1",email = "bodaev2019@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "docker (>=7.1.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
